Move some base classes into __init__
A little more elegant organization of the base classes when there
is a subdirectory named after them. So rules and parsers have the
Rule and Parser base class as part of the __init__ for a slightly
cleaner result.
