+++ precli/renderers/detailed.py
@@ -9,8 +9,8 @@ from rich.table import Table
 from precli.core.level import Level
 from precli.core.metrics import Metrics
 from precli.core.result import Result
-from precli.core.rule import Rule
-from precli.renderers.renderer import Renderer
+from precli.renderers import Renderer
+from precli.rules import Rule
 
 
 class Detailed(Renderer):
