+++ precli/renderers/json.py
@@ -5,8 +5,8 @@ from rich import console
 
 from precli.core.metrics import Metrics
 from precli.core.result import Result
-from precli.core.rule import Rule
-from precli.renderers.renderer import Renderer
+from precli.renderers import Renderer
+from precli.rules import Rule
 
 
 class Json(Renderer):
