+++ precli/rules/python/stdlib/assert/assert.py
@@ -1,7 +1,7 @@
 # Copyright 2023 Secure Saurce LLC
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class Assert(Rule):
