+++ precli/rules/python/stdlib/crypt/crypt_weak_hash.py
@@ -98,7 +98,7 @@ alternatives include ``bcrypt``, ``pbkdf2``, and ``scrypt``.
 """  # noqa: E501
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 WEAK_CRYPT_HASHES = (
