+++ precli/rules/python/stdlib/ftplib/ftp_cleartext.py
@@ -86,7 +86,7 @@ These alternatives include:
 from precli.core.level import Level
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class FtpCleartext(Rule):
