+++ precli/rules/python/stdlib/hmac/hmac_timing_attack.py
@@ -81,7 +81,7 @@ secure alternatives, ``SHA256``, ``SHA-384``, or ``SHA512``.
 from precli.core.level import Level
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 TIMING_VULNERABLE = (
