+++ precli/rules/python/stdlib/hmac/hmac_weak_hash.py
@@ -77,7 +77,7 @@ secure alternatives, ``SHA256``, ``SHA-384``, or ``SHA512``.
 from precli.core.level import Level
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 WEAK_HASHES = ("md4", "md5", "ripemd160", "sha", "sha1")
