+++ precli/rules/python/stdlib/json/json_load.py
@@ -41,7 +41,7 @@ should first sanitize the data to remove any potential malicious code.
 """  # noqa: E501
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class JsonLoad(Rule):
