+++ precli/rules/python/stdlib/marshal/marshal_load.py
@@ -46,7 +46,7 @@ you should first sanitize the data to remove any potential malicious code.
 """  # noqa: E501
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class MarshalLoad(Rule):
