+++ precli/rules/python/stdlib/pickle/pickle_load.py
@@ -58,7 +58,7 @@ to be secure and cannot be used to execute malicious code.
 """  # noqa: E501
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class PickleLoad(Rule):
