+++ precli/rules/python/stdlib/shelve/shelve_open.py
@@ -45,7 +45,7 @@ any potential malicious code.
 """  # noqa: E501
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class ShelveOpen(Rule):
