+++ precli/rules/python/stdlib/ssl/create_unverified_context.py
@@ -57,7 +57,7 @@ these security risks.
 """  # noqa: E501
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 class CreateUnverifiedContext(Rule):
