+++ precli/rules/python/stdlib/ssl/insecure_tls_version.py
@@ -76,7 +76,7 @@ from precli.core.argument import Argument
 from precli.core.level import Level
 from precli.core.location import Location
 from precli.core.result import Result
-from precli.core.rule import Rule
+from precli.rules import Rule
 
 
 INSECURE_VERSIONS = (
