+++ tests/unit/rules/python/stdlib/hmac/test_hmac_timing_attack.py
@@ -2,7 +2,7 @@
 import os
 
 from precli.core.level import Level
-from precli.core.rule import Rule
+from precli.rules import Rule
 from tests.unit.rules.python import test_case
