+++ tests/unit/rules/python/stdlib/telnetlib/test_telnetlib_cleartext.py
@@ -2,7 +2,7 @@
 import os
 
 from precli.core.level import Level
-from precli.core.rule import Rule
+from precli.rules import Rule
 from tests.unit.rules.python import test_case
